requirements.txt
requests
beautifulsoup4
jinja2
python-dotenv
click
pydantic
feedparser
pytest
